# knowsphere-ai-complete_timestamp/knowsphere-ai/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KnowSphere AI API
  version: "6.0.0"
  description: |
    Enterprise Knowledge Intelligence Platform — REST API.

    Scoping note: this specification covers the primary request/response
    shape of every resource group (auth, providers, documents, chat,
    analytics, audit, observability, notifications, health). It documents
    the core, most-used parameters for each endpoint rather than every
    optional query parameter combination across 60+ endpoints — treat this
    as the authoritative contract for integration purposes, and the actual
    route source (`app/*/routes.py`) as the final word on any parameter
    not shown here.

    All endpoints are prefixed with `/api/v1`. Authentication is a JWT
    Bearer token in the `Authorization` header, obtained via `/auth/login`.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /providers:
    get:
      summary: List configured LLM/embedding providers (admin only)
      responses:
        "200": { description: OK }
    post:
      summary: Add a provider (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name, provider_type]
              properties:
                display_name: { type: string }
                provider_type: { type: string, enum: [openai, anthropic, gemini, groq, openrouter, nvidia_nim, ollama, openai_compatible] }
                api_key: { type: string }
                base_url: { type: string }
                capability: { type: string, enum: [llm, embedding, both] }
      responses:
        "201": { description: Provider created }
  /providers/{providerId}/activate:
    post:
      summary: Set this provider as the org-wide default
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Now the default provider }
...
  /observability/langsmith:
    get:
      summary: View LangSmith configuration (key masked, admin only)
      responses:
        "200": { description: OK }
    patch:
      summary: Update LangSmith configuration (admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                api_key: { type: string }
                project_name: { type: string }
                tracing_enabled: { type: boolean }
      responses:
        "200": { description: Updated }
...